src/main/java/com/demo/todo/list/app/security/exception/handling/TodoListAccessDeniedHandler.java
package com.demo.todo.list.app.security.exception.handling;

import com.demo.todo.list.app.model.response.ApiErrorResponse;
import com.fasterxml.jackson.databind.ObjectMapper;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.security.access.AccessDeniedException;
import org.springframework.security.web.access.AccessDeniedHandler;

import java.io.IOException;
import java.util.Date;

@Slf4j
@RequiredArgsConstructor
public class TodoListAccessDeniedHandler implements AccessDeniedHandler {

    private final ObjectMapper objectMapper;

    @Override
    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException {
        log.error("Access denied: {}", accessDeniedException.getMessage());
        response.setStatus(HttpStatus.FORBIDDEN.value());
        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
        ApiErrorResponse apiErrorResponse = ApiErrorResponse.builder()
                .url(request.getRequestURI())
                .createdDate(new Date())
                .build();
        String result = objectMapper.writeValueAsString(apiErrorResponse);
        response.getWriter().write(result);
    }
}
